+++ CHANGELOG.md
@@ -4,6 +4,15 @@
 
 ##### Changed
 * **\[core\]** Remove unused methods `JsUtil.asList(..)` and `JsUtil.getPrototypeForClass(..)`.
+* **\[annotations\]** Add parameter `inject` to the `@ReactComponent` annotation that explicitly controls
+  whether the enhanced component is generated with an `@Inject` annotation on the constructor. The default
+  value for this parameter is `IF_DETECTED` and will generate a constructor using the existing algorithm.
+  i.e. An `@Inject` annotated cosntructor will be generated if a field of method of the component is annotated
+  with `@Inject`.
+* **\[annotations\]** Add parameter `generated` to the `@ReactComponent` annotation that explicitly controls
+  whether a dagger component is created for the react component. The default value for this parameter is
+  `IF_DETECTED` and will generate a constructor using the existing algorithm. i.e. A module is created if the
+  dagger classes are on the compile classpath.
 
 ### [v0.22](https://github.com/react4j/react4j/tree/v0.22) (2017-12-05)
 [Full Changelog](https://github.com/react4j/react4j/compare/v0.21...v0.22)